code/predictionlc50/requirements.txt
asgiref==3.4.1
attrs==21.2.0
cassandra-driver==3.25.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-default-group==1.2.2
cycler==0.11.0
Django==3.2.9
fonttools==4.28.2
geomet==0.2.1.post1
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
Jinja2==3.0.3
joblib==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
missingno==0.5.0
multimethod==1.6
networkx==2.6.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandas-profiling==3.1.0
phik==0.12.0
Pillow==8.4.0
profiling==0.1.3
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
tangled-up-in-unicode==0.1.0
threadpoolctl==3.0.0
tomli==1.2.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
urwid==2.1.2
valuedispatch==0.0.1
visions==0.7.4
waitress==2.0.0
whitenoise==5.3.0
